MOV-8 create a container component for video cards, extract function for mapping arrays with VideoDto -> VideoDtoCard
