+++ frontend/movies-ui/src/app/home/home.component.css
@@ -1,15 +1,3 @@
 :host {
   max-width: var(--max-width);
 }
-
-.videoCardCtn {
-  display: flex;
-  justify-content: center;
-
-  gap: 2rem 1rem;
-  flex-wrap: wrap;
-}
-
-a {
-  all: unset;
-}
