+++ frontend/movies-ui/src/app/home/home.component.html
@@ -1,8 +1,1 @@
-<h2>All videos</h2>
-<div class="videoCardCtn">
-  @for (video of videos; track video.id) {
-    <a [routerLink]="video.id">
-      <app-video-card  [videoDtoCard]="video"/>
-    </a>
-  }
-</div>
+<app-video-box-full [videos]="videos" [heading]="heading"/>
