+++ frontend/movies-ui/src/app/video-box-full/video-box-full.component.css
@@ -1,0 +1,11 @@
+.videoCardCtn {
+  display: flex;
+  justify-content: center;
+
+  gap: 2rem 1rem;
+  flex-wrap: wrap;
+}
+
+a {
+  all: unset;
+}
